Changes files and movies options
Files names changed _ removed and Capital,
Add movie remove movie
