+++ PathiHub/Presentation/UserLogin.cs
@@ -35,7 +35,7 @@ static class UserLogin
             {
                 case "Manager":
                     Console.Clear();
-                    Manager_menu.Start_Menu();
+                    ManagerMenu.StartMenu();
                     Console.WriteLine(acc.GetType());
                     break;
                 case "Financial Manager":
